Refactors history.go and improves performance
