+++ examples/BaSyxHistoryAuditGuardedExample/docker-compose.yml
@@ -28,9 +28,8 @@ services:
       - POSTGRES_CONNMAXLIFETIMEMINUTES=5
       - BASYX_HISTORY_MODE=audit
       - BASYX_HISTORY_RETENTION_DAYS=0
-      - BASYX_HISTORY_FULL_SNAPSHOT_INTERVAL=1
-      - BASYX_HISTORY_IMMUTABILITY=postgres_guarded
       - BASYX_HISTORY_FULL_SNAPSHOT_INTERVAL=5
+      - BASYX_HISTORY_IMMUTABILITY=postgres_guarded
       - BASYX_AUDIT_IDENTITY_MODE=none
       - ABAC_ENABLED=false
       - GENERAL_AASREGISTRYINTEGRATION=true
