+++ internal/common/date_time_test.go
@@ -1,0 +1,68 @@
+/*******************************************************************************
+* Copyright (C) 2026 the Eclipse BaSyx Authors and Fraunhofer IESE
+*
+* Permission is hereby granted, free of charge, to any person obtaining
+* a copy of this software and associated documentation files (the
+* "Software"), to deal in the Software without restriction, including
+* without limitation the rights to use, copy, modify, merge, publish,
+* distribute, sublicense, and/or sell copies of the Software, and to
+* permit persons to whom the Software is furnished to do so, subject to
+* the following conditions:
+*
+* The above copyright notice and this permission notice shall be
+* included in all copies or substantial portions of the Software.
+*
+* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+* EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+* MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+* NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+* LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+* OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+* WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+*
+* SPDX-License-Identifier: MIT
+******************************************************************************/
+
+package common
+
+import (
+	"testing"
+	"time"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestParseISO8601DateTimeAcceptsTimezoneOffset(t *testing.T) {
+	parsed, err := ParseISO8601DateTime("2026-05-28T14:30:00+02:00")
+
+	require.NoError(t, err)
+	require.Equal(t, time.Date(2026, 5, 28, 12, 30, 0, 0, time.UTC), parsed.UTC())
+}
+
+func TestParseISO8601DateTimeAcceptsZuluUTC(t *testing.T) {
+	parsed, err := ParseISO8601DateTime("2026-05-28T12:30:00Z")
+
+	require.NoError(t, err)
+	require.Equal(t, time.Date(2026, 5, 28, 12, 30, 0, 0, time.UTC), parsed.UTC())
+}
+
+func TestParseISO8601DateTimeAcceptsUTCSuffix(t *testing.T) {
+	parsed, err := ParseISO8601DateTime("2026-05-28T12:30:00.123456789 UTC")
+
+	require.NoError(t, err)
+	require.Equal(t, time.Date(2026, 5, 28, 12, 30, 0, 123456789, time.UTC), parsed.UTC())
+}
+
+func TestParseISO8601DateTimeAcceptsDecodedQueryPlusOffset(t *testing.T) {
+	parsed, err := ParseISO8601DateTime("2026-05-28T14:30:00 02:00")
+
+	require.NoError(t, err)
+	require.Equal(t, time.Date(2026, 5, 28, 12, 30, 0, 0, time.UTC), parsed.UTC())
+}
+
+func TestParseISO8601DateTimeAcceptsOffsetWithoutColon(t *testing.T) {
+	parsed, err := ParseISO8601DateTime("2026-05-28T14:30:00+0200")
+
+	require.NoError(t, err)
+	require.Equal(t, time.Date(2026, 5, 28, 12, 30, 0, 0, time.UTC), parsed.UTC())
+}
