+++ pkg/aasxfileserverapi/go/routers.go
@@ -25,6 +25,7 @@ import (
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/chi/v5/middleware"
 
+	"github.com/eclipse-basyx/basyx-go-components/internal/common"
 	model "github.com/eclipse-basyx/basyx-go-components/internal/common/model"
 )
 
@@ -211,10 +212,7 @@ func parseTimes(param string) ([]time.Time, error) {
 
 // parseTime will parses a string parameter into a time.Time using the RFC3339 format
 func parseTime(param string) (time.Time, error) {
-	if param == "" {
-		return time.Time{}, nil
-	}
-	return time.Parse(time.RFC3339, param)
+	return common.ParseISO8601DateTime(param)
 }
 
 type Number interface {
