+++ pkg/smregistry/helpers.go
@@ -115,10 +115,7 @@ func EncodeJSONResponse(i interface{}, status *int, w http.ResponseWriter) error
 
 // parseTime will parses a string parameter into a time.Time using the RFC3339 format
 func parseTime(param string) (time.Time, error) {
-	if param == "" {
-		return time.Time{}, nil
-	}
-	return time.Parse(time.RFC3339, param)
+	return common.ParseISO8601DateTime(param)
 }
 
 type Number interface {
